fix some error at transactions and delete inactive game

+++ dominoapp/management/commands/delete_inactive_game.py
@@ -2,21 +2,19 @@ from django.core.management.base import BaseCommand
 from django.utils.timezone import timedelta, now
 from django.db.models import Q
 from dominoapp.models import DominoGame
+from dominoapp.views import playersCount
 
 class Command(BaseCommand):
     help = "Delete games with an inactivity time greater than 24 hours."
 
     def handle(self, *args, **options):
-        expired_time = now() - timedelta(hours = 24)
+        expired_time = now() - timedelta(hours = 15)
         
         games_models = DominoGame.objects.filter(
-            lastTime1__lt=expired_time,
-            lastTime2__lt=expired_time,
-            lastTime3__lt=expired_time,
-            lastTime4__lt=expired_time
+            start_time__lt=expired_time
             ).exclude(id__in=[33,36,55])
 
         for game in games_models:
-            game.delete()
-
+            if playersCount(game) == 0:
+                game.delete()
         return
